rename header file and wrap logo in link
